Added initial code
The code can sample the coverage time for a Johnson-Mehl tessellation in $[0,1]^2$ (with given arrival rate). 
It can also plot a diagram with the empirical distribution, the new limit expression of Penrose for this distribution, and the predicted distribution from Chiu (1995).

The predicted distribution is computed using Chiu's limiting CDF for the coverage time as if it were true for finite rho rather than asymptotically. I did this because Chiu and Penrose use different normalisations of the coverage time to get a convergent quantity. It does have the unfortunate side-effect that Chiu's prediction depends on the arrival rate $rho$. I suppose I could also have converted them both into direct statements about the distribution of the coverage time, maybe I'll do that later.