Everything except get_ball_pixels works for d>=3
get_ball_pixels is the difficult and possibly slow bit.
The CNRM team only wrote their code for d=2 and d=3, so maybe they tried general d and didn't find a fast method. I'm not that concerned about perfectly-optimal performance though, so I don't mind implementing a not-so-quick method.
I think I'll use the recursive method which is described as a comment in jmutils.py
(I also need to test the implemented stuff: there's a chance Numba won't like it being in d dimensions.)